Track block damage duration
This is now used to track Furniture breaking with a custom breaking duration and can be applied to other custom blocks in the future.

This also replaces spam clicking when using the perks of the Hasty Pickaxe which allow breaking of otherwise unbreakable blocks, namely Bedrock and End Portal Frames.